Fix slow transcript data retrieval (exp overview)
The query used to retrieve transcript data for the table displayed on
the experiment overview was very slow (not using the 'experiment_id'
index, apparently because of the use of ORDER BY/LIMIT). The query is
now using the correct index (modification of 'ORDER BY' and 'WHERE'
clauses). This change should fix #5.

An alternative way to fix the problem is to use MySQL index hints (more
details in issue #5). Some first attempt to do that can be found in the
'Transcripts' model (commented).